Allow user to log out
This change adds the feature allowing users to log out. It also includes
fairly extensive refactoring to make this feature easier to add,
including the following:

- Authenticate.js is simplified to only receive functions to be applied
to buttons.

- GoalSetter.js got more logic added to it but always renders the same
components (with different props).

- GoalsForm.js is now sensitive to whether there is a token.

I'm not at all convinced that this is a good final organization for the
code but it works for now. What I'd really like is some sort of separate
"before hook" to check if the user is authenticated but I don't really
know what that looks like in React and I'm just using the presence or
absence of a token as proxy for whether the user is logged in.

Fixes #14