Implement adding issues
This change allows the user to add one issue associated with the
milestone they are also adding. The state of GoalsForm.js has been
reworked two include two nested objects, one for the milestone and one for the
issue. Field components (MilestoneField and IssueField) have also been
extracted.

This change also removes the display of current milestones (but the API
function remains in the GitHubCommunicator since it might be useful
later).

Fixes #21