Create pull request with markdown file
This commit adds a textbox to the form, in which the user can edit a
pre-populated goals markdown file. Upon submission, the markdown file
will be made into a pull request on their repo so that they can confirm
that they want to add it to their project.

Creating a commit with the GitHub API is a pretty involved process so
I've tucked it all away in the pullRequester.js file.

Fixes #2